test: merge redundant e2e specs into existing builds
